// Fontend/TNT.N8.Web/TNT.N8.Web/ClientApp/src/app/shared/models/VendorMappingOption.model.ts
export class VendorMappingOption {
    id: string;
    vendorId: string;
    optionId: string;
    createdById: string;
    createdDate: string;
    updatedById: string | null;
    updatedDate: string | null;
    price: number | null;
    soLuongToiThieu: number | null;
    yeuCauThanhToan: number | null;
    giaTriThanhToan: number | null;
    thoiGianTu: Date | null;
    thoiGianDen: Date | null;
    stt: number | null;

    email: number | null;
    address: number | null;
    phoneNumber: number | null;
    vendorName: number | null;
    vendorGroupName: number | null;
    vendorCode: number | null;

    chietKhauId: number | null;
    giaTriChietKhau: number | null;
    donViTienId: string | null;
    thueGtgt: number | null;

}